Consolidate seed scripts to use only seed:exercises:wrkout
- Moved upsertExercises function from seedExercises.ts to seedExercisesFromWrkout.ts
- Removed old CSV-based seed:exercises script from package.json
- Deleted seedExercises.ts, seedExercises.test.ts, and seed_data directory
- Updated imports in tests to reference the consolidated script
- Kept only seed:exercises:wrkout as the single seeding method
