--- backend/src/seed_data/exercises.csv
@@ -1,64 +1,0 @@
-slug,name,tags
-arm-circles,Arm Circles,warmup;mobility;shoulders
-leg-swings-forward-back,Leg Swings (Forward/Back),warmup;mobility;lower-body
-leg-swings-side-side,Leg Swings (Side to Side),warmup;mobility;lower-body
-cat-cow,Cat-Cow,warmup;mobility;spine
-ankle-circles,Ankle Circles,warmup;mobility;lower-body
-hip-cars,Hip CARs,warmup;mobility;hips
-worlds-greatest-stretch,World's Greatest Stretch,warmup;mobility;full-body
-deep-squat-pry,Deep Squat Pry,warmup;mobility;hips
-scapular-pushup,Scapular Push-Up,warmup;mobility;shoulders
-thoracic-rotation-open-book,Open Book Rotation,warmup;mobility;spine
-hamstring-stretch-standing,Standing Hamstring Stretch,stretch;lower-body
-quad-stretch-standing,Standing Quad Stretch,stretch;lower-body
-calf-stretch-wall,Calf Stretch (Wall),stretch;lower-body
-hip-flexor-stretch-kneeling,Kneeling Hip Flexor Stretch,stretch;lower-body
-adductor-stretch-side-lean,Side Adductor Stretch,stretch;lower-body
-piriformis-stretch,Seated Piriformis Stretch,stretch;lower-body
-childs-pose,Child's Pose,stretch;spine
-upper-trap-stretch,Upper Trap Stretch,stretch;upper-body
-lat-stretch-overhead,Overhead Lat Stretch,stretch;upper-body
-thoracic-extension-foam-roller,Thoracic Extension on Foam Roller,stretch;spine
-barbell-squat,Barbell Squat,strength;lower-body;barbell
-goblet-squat,Goblet Squat,strength;lower-body;dumbbell
-barbell-deadlift,Barbell Deadlift,strength;lower-body;barbell
-romanian-deadlift-barbell,Barbell Romanian Deadlift,strength;lower-body;barbell
-trap-bar-deadlift,Trap Bar Deadlift,strength;lower-body;trap-bar
-barbell-reverse-lunge,Barbell Reverse Lunge,strength;lower-body;barbell
-dumbbell-lunge,Dumbbell Lunge,strength;lower-body;dumbbell
-reverse-walking-lunge,Reverse Walking Lunge,strength;lower-body;dumbbell
-walking-lunge,Walking Lunge,strength;lower-body;dumbbell
-bulgarian-split-squat-dumbbell,Dumbbell Bulgarian Split Squat,strength;lower-body;dumbbell
-cossack-squat,Cossack Squat,strength;lower-body;mobility
-step-up-dumbbell,Dumbbell Step-Up,strength;lower-body;dumbbell
-calf-raise-dumbbell,Dumbbell Calf Raise,strength;lower-body;dumbbell
-barbell-hip-thrust,Barbell Hip Thrust,strength;lower-body;barbell
-single-leg-rdl-dumbbell,Single-Leg RDL (Dumbbell),strength;lower-body;dumbbell
-barbell-overhead-press,Barbell Overhead Press,strength;upper-body;barbell
-dumbbell-bench-press,Dumbbell Bench Press,strength;upper-body;dumbbell
-push-up,Push-Up,strength;upper-body;bodyweight
-pull-up,Pull-Up,strength;upper-body;bodyweight
-dumbbell-row,Dumbbell Row,strength;upper-body;dumbbell
-landmine-press-single-arm,Single Arm Landmine Press,strength;upper-body;landmine
-medicine-ball-slam,Medicine Ball Slam,strength;full-body;power
-medicine-ball-rotational-throw,Rotational Med Ball Throw,strength;core;power
-stationary-bike,Stationary Bike,cardio;machine
-treadmill-run,Treadmill Run,cardio;machine
-rowing-machine,Rowing Machine,cardio;machine
-elliptical,Elliptical,cardio;machine
-jump-rope,Jump Rope,cardio;bodyweight
-shuttle-runs,Shuttle Runs,cardio;agility;field
-high-knees,High Knees,cardio;bodyweight
-butt-kicks,Butt Kicks,cardio;bodyweight
-skips-a,Skipping A-March,cardio;mobility;drill
-skips-b,Skipping B-March,cardio;mobility;drill
-pallof-press-band,Pallof Press (Band),band;core;anti-rotation
-band-pull-aparts,Band Pull-Aparts,band;upper-body;shoulders
-tke-band,Terminal Knee Extension (Band),band;lower-body;knee
-lateral-band-walks,Lateral Band Walks,band;lower-body;glutes
-monster-walks-band,Monster Walks (Band),band;lower-body;glutes
-plank,Plank,core;bodyweight;strength
-side-plank,Side Plank,core;bodyweight;strength
-hollow-hold,Hollow Hold,core;bodyweight
-reverse-crunch,Reverse Crunch,core;bodyweight
-mountain-climbers,Mountain Climbers,core;cardio;bodyweight
